# openspec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AstrBot OpenAPI v1
  version: 0.1.0
  description: >
    Target REST contract for migrating AstrBot HTTP APIs to /api/v1.
    Dynamic AstrBot configuration payloads are intentionally modeled as open
    JSON objects because their schemas are provided at runtime by template
    endpoints.
    Developer API keys currently support these scopes only: bot, provider,
    persona, im, config, chat, data, file, plugin, mcp, skill. The config scope also
    grants bot and provider access.
servers:
  - url: http://localhost:6185
    description: Local AstrBot server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/knowledge-bases:
    get:
      tags: [Knowledge Base]
      summary: List knowledge bases
      operationId: listKnowledgeBases
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: refresh_stats
          in: query
          schema:
            type: boolean
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [Knowledge Base]
      summary: Create a knowledge base
      operationId: createKnowledgeBase
      x-astrbot-scope: kb
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeBaseRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}:
    get:
      tags: [Knowledge Base]
      summary: Get a knowledge base
      operationId: getKnowledgeBase
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [Knowledge Base]
      summary: Update a knowledge base
      operationId: updateKnowledgeBase
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeBaseRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Knowledge Base]
      summary: Delete a knowledge base
      operationId: deleteKnowledgeBase
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/stats:
    get:
      tags: [Knowledge Base]
      summary: Get knowledge base stats
      operationId: getKnowledgeBaseStats
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/documents:
    get:
      tags: [Knowledge Base]
      summary: List knowledge base documents
      operationId: listKnowledgeDocuments
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [Knowledge Base]
      summary: Upload a document to a knowledge base
      operationId: uploadKnowledgeDocument
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/KnowledgeDocumentUploadRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/documents/import:
    post:
      tags: [Knowledge Base]
      summary: Import documents already available on the server
      operationId: importKnowledgeDocuments
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeDocumentImportRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/documents/import-url:
    post:
      tags: [Knowledge Base]
      summary: Import a document from URL
      operationId: importKnowledgeDocumentFromUrl
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeDocumentUrlImportRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/documents/{document_id}:
    get:
      tags: [Knowledge Base]
      summary: Get a knowledge base document
      operationId: getKnowledgeDocument
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Knowledge Base]
      summary: Delete a knowledge base document
      operationId: deleteKnowledgeDocument
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/chunks:
    get:
      tags: [Knowledge Base]
      summary: List document chunks
      operationId: listKnowledgeChunks
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
        - name: document_id
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/chunks/{chunk_id}:
    delete:
      tags: [Knowledge Base]
      summary: Delete a document chunk
      operationId: deleteKnowledgeChunk
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
        - $ref: "#/components/parameters/ChunkId"
        - name: document_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/{kb_id}/retrieve:
    post:
      tags: [Knowledge Base]
      summary: Retrieve knowledge base chunks
      operationId: retrieveKnowledgeBase
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeRetrieveRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/knowledge-bases/tasks/{task_id}:
    get:
      tags: [Knowledge Base]
      summary: Get knowledge base import task progress
      operationId: getKnowledgeTask
      x-astrbot-scope: kb
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
  /api/v1/stats:
    get:
      tags: [Stats]
      summary: Get runtime statistics
      operationId: getStats
      x-astrbot-scope: system
      parameters:
        - name: offset_sec
          in: query
          schema:
            type: integer
            default: 86400
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Standard AstrBot success response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
...
  parameters:
...
    ChunkId:
      name: chunk_id
      in: path
      required: true
      schema:
        type: string
...
    DocumentId:
      name: document_id
      in: path
      required: true
      schema:
        type: string
...
    KbId:
      name: kb_id
      in: path
      required: true
      schema:
        type: string
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
...
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          const: ok
        message:
          type: string
        data: {}
      additionalProperties: false
...
    DynamicConfig:
      type: object
      additionalProperties: true
...
    KnowledgeBaseRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        embedding_provider_id:
          type: string
        rerank_provider_id:
          type: string
        chunking:
          $ref: "#/components/schemas/DynamicConfig"
        metadata:
          $ref: "#/components/schemas/DynamicConfig"
      additionalProperties: false
    KnowledgeDocumentUploadRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
        parser:
          type: string
    KnowledgeDocumentImportRequest:
      type: object
      required: [documents]
      properties:
        documents:
          type: array
          items:
            type: object
            required: [file_name, chunks]
            properties:
              file_name:
                type: string
              chunks:
                type: array
                items:
                  type: string
            additionalProperties: true
        batch_size:
          type: integer
        tasks_limit:
          type: integer
        max_retries:
          type: integer
      additionalProperties: false
    KnowledgeDocumentUrlImportRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        chunk_size:
          type: integer
        chunk_overlap:
          type: integer
        batch_size:
          type: integer
        tasks_limit:
          type: integer
        max_retries:
          type: integer
        enable_cleaning:
          type: boolean
        cleaning_provider_id:
          type: [string, "null"]
      additionalProperties: false
    KnowledgeRetrieveRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
        top_k:
          type: integer
          default: 5
        score_threshold:
          type: number
      additionalProperties: false
...